# fix: avoid copying label values from tsdb unless required