**What this PR does / why we need it**:
We are seeing OOMKills on index gateways due to a huge amount of allocations from [LabelValues](https://github.com/grafana/loki/blob/5625464332fc0b3371b744d60d36acaf2600665a/pkg/storage/stores/shipper/indexshipper/tsdb/index/index.go#L1631) method of tsdb index reader. It happens due to copying all the label values read from the tsdb. [This problem arises when the label matchers have any negative matchers](https://github.com/grafana/loki/blob/5625464332fc0b3371b744d60d36acaf2600665a/pkg/storage/stores/shipper/indexshipper/tsdb/querier.go#L175-L191).

I have changed the code to avoid making a copy of the label values unless we want to send it back as a response. I have checked all the usages of that method, and the only place we need to copy the values is when the `LabelValues` method is called explicitly to get all the values.

Here is the benchmark showing the difference in performance:
```
benchmark                              old ns/op      new ns/op     delta
BenchmarkTSDBIndex_GetChunkRefs-10     875881104      927752604     +5.92%
BenchmarkTSDBIndex_GetChunkRefs-10     879311146      862689354     -1.89%
BenchmarkTSDBIndex_GetChunkRefs-10     881187479      926149750     +5.10%
BenchmarkTSDBIndex_GetChunkRefs-10     910079084      899049584     -1.21%
BenchmarkTSDBIndex_GetChunkRefs-10     1014711583     911985770     -10.12%

benchmark                              old allocs     new allocs     delta
BenchmarkTSDBIndex_GetChunkRefs-10     2000140        1000154        -50.00%
BenchmarkTSDBIndex_GetChunkRefs-10     2000139        1000151        -50.00%
BenchmarkTSDBIndex_GetChunkRefs-10     2000141        1000150        -50.00%
BenchmarkTSDBIndex_GetChunkRefs-10     2000141        1000149        -50.00%
BenchmarkTSDBIndex_GetChunkRefs-10     2000146        1000153        -50.00%

benchmark                              old bytes     new bytes     delta
BenchmarkTSDBIndex_GetChunkRefs-10     176039840     168119396     -4.50%
BenchmarkTSDBIndex_GetChunkRefs-10     176039752     168118520     -4.50%
BenchmarkTSDBIndex_GetChunkRefs-10     176039944     168118504     -4.50%
BenchmarkTSDBIndex_GetChunkRefs-10     176039952     168118408     -4.50%
BenchmarkTSDBIndex_GetChunkRefs-10     176057848     168119296     -4.51%
```

**Special notes for your reviewer**:
I introduced the copying of label values in PR https://github.com/grafana/loki/pull/7502 to avoid corruption of values when the underlying tsdb is closed.